changes
git-svn-id: svn://localhost/public/MoSync/trunk@25 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ tools/RuntimeBuilder/BuildJavaME.rb
@@ -20,10 +20,6 @@ require 'ftools'
 require 'fileutils'
 			
 class RuntimeBuilder 
-
-	def javaSource
-		@javaSource = "../../runtimes/java";
-	end
 	
 	def javame(runtime_dir)
 		javameBuilder(runtime_dir, false);
@@ -55,7 +51,7 @@ class RuntimeBuilder
 #		String sharedSrcDir = javaSourceDir + "shared\\";
 		sharedSrc = $settings[:java_source] + "\\shared";
 #		File f = new File(sharedSrcDir);
-
+		Dir.foreach("testdir") {|x| puts "Got #{x}" }
 #		String files[] = f.list();
 #		for (int i = 0; i < files.length; i++)
 #		{
@@ -96,7 +92,9 @@ class RuntimeBuilder
 		if File.exist? configBakFile
 			File.delete configBakFile
 		end
-		File.copy(configFile, configBakFile)
+		if File.exist? configFile
+			File.copy(configFile, configBakFile)
+		end
 		File.copy( runtime_dir + "config.h", configFile)
 		
 		# Preprocess all the shared java files and store result in temporary location
@@ -105,18 +103,67 @@ class RuntimeBuilder
 		# Preprocess all the platform dependant java files and store result in temporary location
 		
 		# Restore config_platform.h
-		File.copy(configBakFile, configFile)
-		File.delete configBakFile
+		if File.exist? configBakFile
+			File.copy(configBakFile, configFile)
+			File.delete configBakFile
+		end	
 	
 		# Compile Java source
-		
+#		exec("javac -source 1.4 -target 1.4 " + " -d " + classDir + " -classpath " + classDir +
+#			" -bootclasspath " +
+#			javaMESdkDir + "j2melib\\jsr082.jar;" +
+#			javaMESdkDir + "j2melib\\cldcapi11.jar;" +
+#			javaMESdkDir + "j2melib\\midpapi20.jar;" +
+#			javaMESdkDir + "j2melib\\wma20.jar;" +
+#			javaMESdkDir + "j2melib\\jsr179.jar " + " " +
+#			tempDir + "*.java", true);
+			
 		# Generate Manifest file
+#		DataOutputStream dos = new DataOutputStream(new FileOutputStream(tempDir + "\\manifest.mf"));
+#		writeDosCompatibleString(dos, "MIDlet-1: MoSyncRuntime" + vmInfo.number +
+#			" , MoSyncRuntime" + vmInfo.number + ".png, MAMidlet\n");
+#		writeDosCompatibleString(dos, "MIDlet-Name: MAMidlet\n");
+#		writeDosCompatibleString(dos, "MIDlet-Vendor: Mobile Sorcery\n");
+#		writeDosCompatibleString(dos, "MIDlet-Version: 1.0\n");
+#		writeDosCompatibleString(dos, "MicroEdition-Configuration: CLDC-1.1\n");
+#		writeDosCompatibleString(dos, "MicroEdition-Profile: MIDP-2.0\n");
+#		dos.close();
 		
 		# Build jar file
+#		exec("jar cfm " +
+#		runtimeDir + "MoSyncRuntimeTemp.jar " +
+#		tempDir + "manifest.mf " +
+#		"-C " + classDir + " .", true);
 		
 		# Obfuscate java binaries
+#		exec("java " +
+#		"-jar " + javaMESdkDir + "bin\\proguard.jar " +
+#		"-injars " + runtimeDir + "MoSyncRuntimeTemp.jar " +
+#		"-libraryjars " + javaMESdkDir + "j2melib\\cldcapi11.jar " +
+#		"-libraryjars " + javaMESdkDir + "j2melib\\midpapi20.jar " +
+#		"-libraryjars " + javaMESdkDir + "j2melib\\jsr082.jar " +
+#		"-libraryjars " + javaMESdkDir + "j2melib\\jsr179.jar " +
+#		"-libraryjars " + javaMESdkDir + "j2melib\\wma20.jar " +
+#		"-dontusemixedcaseclassnames " +
+#		"-outjars " + runtimeDir + "MoSyncRuntimeObfuscated.jar " +
+#		"-keep public class MAMidlet", true);
+		
+#		String D = debug ? "D" : "";
+#		exec("move " + runtimeDir + "MoSyncRuntimeObfuscated.jar " + runtimeDir +
+#			"MoSyncRuntime" + D + ".jar ", true);
 		
 		# Preverify java binaries
+#		exec(javaMESdkDir + "bin\\" + "preverify " +
+#			//				"-d " + preverifiedClassDir + " " +
+#		"-d " + runtimeDir.substring(0, runtimeDir.length() - 1) + " " +
+#		"-classpath " +
+#			//				runtimeDir + "MoSyncRuntimeObfuscated.jar" +
+#		javaMESdkDir + "j2melib\\jsr082.jar;" +
+#		javaMESdkDir + "j2melib\\cldcapi11.jar;" +
+#		javaMESdkDir + "j2melib\\midpapi20.jar;" +
+#		javaMESdkDir + "j2melib\\jsr179.jar;" +
+#		javaMESdkDir + "j2melib\\wma20.jar " +
+#		runtimeDir + "MoSyncRuntime" + D + ".jar", true);
 		
 		# Clean and delete all the temporary folders
 		FileUtils.rm_rf temp_dir
